ResumeMatchPro-telegram-alert/pyproject.toml
[project]
name = "resumematchpro-telegram-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.22.1",
    "requests>=2.32.3",
]
